+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Fixed exception handling in web requests
+- Internal fixes for environment variables
 
 ## [4.2.0] - 2023-04-16
 ### Added